Fix a bug in answer extraction from question statements
The original version used a regex to try to find the interior
of the LaTex `\boxed{}` but didn't do so recursively so failed
to find nested braces like `\boxed{\frac{1}{2}}`.  We fixed
this by removing the regex and just iterating through the
string until we found the closing braces.  A recurrsive
regex solution would likely have been possible but we chose to
keep it simple.